Remove the separate folder for enums
