--- client-dotnet/src/Controller.Api/enum/ContentType.cs
@@ -1,7 +1,0 @@
-namespace Controller.Api
-{
-  public enum ContentType {
-      JSON,
-      X_WWW_FORM_URLENCODED
-    }
-}
