--- client-dotnet/src/Controller.Api/enum/HttpVerb.cs
@@ -1,6 +1,0 @@
-namespace Controller.Api {
-  public enum HttpVerb {
-    GET,
-    POST
-  }
-}
