--- client-dotnet/src/Controller.Api/enum/Type.cs
@@ -1,11 +1,0 @@
-namespace Controller.Api
-{
-  /**
-   * The type of extra heuristic.
-   * Note: for the moment, we only have heuristics on SQL commands
-   */
-  public enum Type {
-    SQL
-  }
-
-}
